+++ NEWS.md
@@ -6,6 +6,7 @@
 
 ## Additions
 
+- Each agent can be assigned a different dose-response function (by assigning a vector of functions to `fun` in `mbnma.run()`) so that multiple functions can be modelled simultaneously. Some downstream package functions still may not yet work with these models though.
 - `mbnma.network` objects returned from `plot.mbnma.network` now have specific igraph attributes assigned to them, which can be easily changed by the user.
 
 ## Bug fixes
@@ -14,6 +15,9 @@
 - Exponential function models were not working previously but the dose-response function has been rewritten so that it runs the model correctly.
 - DIC reported correctly in output when using plugin (`pd="plugin"`), or Kullback-Leibler divergence (`pd="pd.kl"`)
 - Using the argument `parallel=TRUE` in `mbnma.run()` (or wrapper functions) now properly runs JAGS in parallel on multiple cores.
+
+### Minor
+- Downstream mbnma-related objects now store `mbnma.network` in their output rather than just treatment and agent names.
 
 
 # MBNMAdose 0.2.4